+++ src/steadysun/__init__.py
@@ -1,10 +1,10 @@
 from importlib.metadata import PackageNotFoundError, version
 
-from . import exceptions, forecast, pvsystem
+from . import forecast, pvsystem, steadysun_api
 
 try:
     __version__ = version("steadysun")
 except PackageNotFoundError:
     __version__ = "unknown version"
 
-__all__ = ["exceptions", "forecast", "pvsystem"]
+__all__ = ["forecast", "pvsystem", "steadysun_api"]