+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Change Log
 
+## [0.0.9](https://test.pypi.org/project/steadysun/0.0.9) (2024-10-04)
+
+1st "good" release of the steadysun package ! >ith too much changes
+
+- **ADD** PVSystem model fully created [steadysun.pvsystem](src/steadysun/pvsystem)
+- **UPDATE** [steadysun.forecast](src/steadysun/forecast.py)
+- **ADD** Handling of Api errors (HttpError) with custom message in [steadysun._api](src/steadysun/_api.py)
+- **ADD** Tests for all functions
+
 ## [0.0.8](https://test.pypi.org/project/steadysun/0.0.8) (2024-09-29)
 
 8th test release of the steadysun package.